pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "dspy>=2.6.27",
    "fastapi>=0.116.1",
    "marimo>=0.14.16",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "python-json-logger>=3.3.0",
    "requests>=2.32.4",
    "ruff>=0.12.8",
    "uvicorn>=0.35.0",
    "yoyo-migrations>=9.0.0",
]
